Update to do component if user click on delete button in edit component delete that to do

+++ src/app/components/to-do/to-do.component.ts
@@ -14,12 +14,12 @@ import { AggreementComponent } from '../aggreement/aggreement.component';
   styleUrls: ['./to-do.component.css'],
 })
 export class ToDoComponent implements OnInit {
-  @Input('toDo') toDo: ToDo = new ToDo;
-  @Output("getToDos") getToDos = new EventEmitter()
+  @Input('toDo') toDo: ToDo = new ToDo();
+  @Output('getToDos') getToDos = new EventEmitter();
   displayDatePicker: boolean = false;
 
-  getAllToDos(){
-    this.getToDos.next('')
+  getAllToDos() {
+    this.getToDos.next('');
   }
 
   constructor(
@@ -42,23 +42,43 @@ export class ToDoComponent implements OnInit {
     );
   }
 
-  openAggreementDialog(){
-    this.dialog.open(AggreementComponent, {data: {title: "Are you sure you want to delete this to do"}})
-    .afterClosed().subscribe((result:any) => {if (result === "Yes") {
-      this.deleteToDo()
-    }})
+  openAggreementDialog() {
+    this.dialog
+      .open(AggreementComponent, {
+        data: { title: 'Are you sure you want to delete this to do' },
+      })
+      .afterClosed()
+      .subscribe((result: any) => {
+        if (result === 'Yes') {
+          this.deleteToDo();
+        }
+      });
   }
 
-  deleteToDo(){
-    this.toDoService.delete("to-do/delete-to-do/" + localStorage.getItem('folder') + "/" + localStorage.getItem("list") + "/" + localStorage.getItem("username") + "/" + this.toDo.id).subscribe(
-      (response: any) => {
-        this.notifier.notify(NotificationType.SUCCESS, "The to do successfully deleted")
-        this.getAllToDos();
-      },
-      (error: HttpErrorResponse) => {
-        console.log(error);
-      }
-    )
+  deleteToDo() {
+    this.toDoService
+      .delete(
+        'to-do/delete-to-do/' +
+          localStorage.getItem('folder') +
+          '/' +
+          localStorage.getItem('list') +
+          '/' +
+          localStorage.getItem('username') +
+          '/' +
+          this.toDo.id
+      )
+      .subscribe(
+        (response: any) => {
+          this.notifier.notify(
+            NotificationType.SUCCESS,
+            'The to do successfully deleted'
+          );
+          this.getAllToDos();
+        },
+        (error: HttpErrorResponse) => {
+          console.log(error);
+        }
+      );
   }
 
   updateToDoBody(event: any) {
@@ -67,7 +87,7 @@ export class ToDoComponent implements OnInit {
       this.toDoService.update('to-do/update-to-do', this.toDo).subscribe(
         (response: any) => {
           this.notifier.notify(NotificationType.SUCCESS, 'Success');
-          this.getAllToDos()
+          this.getAllToDos();
         },
         (error: HttpErrorResponse) => {
           this.notifier.notify(NotificationType.ERROR, error.message);
@@ -76,8 +96,15 @@ export class ToDoComponent implements OnInit {
     }
   }
 
-  openEditToDoDialog(){
-    this.dialog.open(EditToDoComponent,{data: {todo: this.toDo}})
+  openEditToDoDialog() {
+    this.dialog
+      .open(EditToDoComponent, { data: { todo: this.toDo } })
+      .afterClosed()
+      .subscribe((result) => {
+        if (result === 'delete') {
+          this.deleteToDo()
+        }
+      });
   }
 
   changeDisplayOfDatePicker() {
